+++ changelog.md
@@ -25,3 +25,7 @@
 
 #1.0.5
 * Fixed wrong ref to SCSS
+
+#1.0.6
+* Added optional forRoot method to be used in your root Module to pass options (see ReadMe)
+* Fixed readme wrong reference: NgxTourWizardModule instead of TourWizardModule